cdk/requirements.txt
annotated-types==0.5.0
attrs==21.4.0
aws-cdk-lib==2.176.0
aws-cdk.asset-awscli-v1==2.2.220
aws-cdk.asset-kubectl-v20==2.1.3
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.aws-apigatewayv2-alpha==2.3.0a0
aws-cdk.aws-apigatewayv2-integrations-alpha==2.3.0a0
aws-cdk.cloud-assembly-schema==39.2.4
cattrs==22.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
constructs==10.1.16
docker==6.1.3
exceptiongroup==1.0.0rc7
idna==3.4
importlib-resources==6.0.1
jsii==1.106.0
packaging==23.1
publication==0.0.3
pydantic==2.1.1
pydantic-settings==2.0.2
pydantic_core==2.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.7.1
urllib3==2.0.4
websocket-client==1.6.1
zipp==3.16.2
